requirements.txt
ailment==9.2.117
angr==9.2.117
archinfo==9.2.117
asttokens==2.4.1
bcrypt==4.1.3
bitarray==2.9.2
bitstring==4.2.3
cachetools==5.5.0
capstone==5.0.2
cart==1.2.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
claripy==9.2.117
cle==9.2.117
colored-traceback==0.3.0
CppHeaderParser==2.7.4
cryptography==42.0.8
decorator==5.1.1
dpkt==1.9.8
executing==2.0.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
intervaltree==3.1.0
ipython==8.27.0
itanium-demangler==1.1
jedi==0.19.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
monkeyhex==1.7.4
mpmath==1.3.0
mulpyplexer==0.9
nampa==0.1.1
networkx==3.3
numpy==2.1.0
packaging==24.1
paramiko==3.4.0
parso==0.8.4
pefile==2024.8.26
pexpect==4.9.0
plumbum==1.8.3
ply==3.11
prompt_toolkit==3.0.47
protobuf==5.28.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pwntools==4.12.0
pycparser==2.22
pycryptodome==3.20.0
pydot==3.0.1
pyelftools==0.31
pyformlang==1.0.10
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.4
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
pyvex==9.2.117
requests==2.32.3
rich==13.8.0
ROPGadget==7.4
rpyc==6.0.0
scapy==2.5.0
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
stack-data==0.6.3
sympy==1.13.2
traitlets==5.14.3
typing_extensions==4.12.2
unicorn==2.0.1.post1
unique_log_filter==0.1.0
unix-ar==0.2.1
urllib3==2.2.2
wcwidth==0.2.13
z3-solver==4.13.0.0
zstandard==0.22.0
